Adds service-level group/version settings for Stargate protocol.
Adds an extra() key-value array for NamingOptions which can be use to
customizing NamingService. Group and version on NamingOptions are
deprecated now.

StargateZookeeperNamingService will use group and version in extra and
fallback to deprecated group and version on NamingOptions for service
registration. Extra options other than `group`, `version` and
`interface` will be encoded in registration URI as well.